Move complexity to own file
